[CLI] Move everything around again
