+++ src/commands/new.js
@@ -1,10 +1,11 @@
 import path from 'path';
 import {readFileSync, writeFileSync} from 'fs';
-import exists from '../util/exists';
 import {sync as mkdirpSync} from 'mkdirp';
 
 import {Promise} from 'es6-promise';
 import recursive from 'recursive-readdir';
+
+import exists from '../util/exists';
 
 const templateDir = path.join(__dirname, '../../template');
